+++ src/main/java/umc/meme/auth/domain/user/entity/UserRepository.java
@@ -1,10 +1,7 @@
-package umc.meme.auth.domain.user.domain;
+package umc.meme.auth.domain.user.entity;
 
 import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.data.jpa.repository.Modifying;
-import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.query.Param;
-import org.springframework.transaction.annotation.Transactional;
 
 import java.util.Optional;
