+++ src/main/java/umc/meme/auth/global/auth/AuthService.java
@@ -9,10 +9,10 @@ import org.springframework.security.core.userdetails.UserDetails;
 import org.springframework.security.core.userdetails.UsernameNotFoundException;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
-import umc.meme.auth.domain.token.domain.Token;
-import umc.meme.auth.domain.token.domain.TokenRepository;
-import umc.meme.auth.domain.user.domain.User;
-import umc.meme.auth.domain.user.domain.UserRepository;
+import umc.meme.auth.domain.token.entity.Token;
+import umc.meme.auth.domain.token.entity.TokenRepository;
+import umc.meme.auth.domain.user.entity.User;
+import umc.meme.auth.domain.user.entity.UserRepository;
 import umc.meme.auth.global.auth.dto.AuthRequest;
 import umc.meme.auth.global.auth.dto.AuthResponse;
 import umc.meme.auth.global.exception.handler.JwtHandler;
